feat: go 1.22 and slog migration
