+++ .github/workflows/release.yml
@@ -18,7 +18,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        go: ['1.19', '1.20', '1.21']
+        go: ['1.21', '1.22']
     name: Go ${{ matrix.go }} test
     steps:
       - uses: actions/checkout@v4
