+++ README.md
@@ -115,10 +115,9 @@ Versions that also build are marked with :warning:.
 
 | Version | Supported          |
 | ------- | ------------------ |
-| <1.19   | :x:                |
-| 1.19    | :warning:          |
-| 1.20    | :white_check_mark: |
+| <1.21   | :x:                |
 | 1.21    | :white_check_mark: |
+| 1.22    | :white_check_mark: |
 
 ## Why another library
