+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/zitadel/oidc/v3
 
-go 1.19
+go 1.21
 
 require (
 	github.com/bmatcuk/doublestar/v4 v4.6.1
@@ -16,11 +16,10 @@ require (
 	github.com/rs/cors v1.10.1
 	github.com/sirupsen/logrus v1.9.3
 	github.com/stretchr/testify v1.8.4
-	github.com/zitadel/logging v0.5.0
+	github.com/zitadel/logging v0.5.1-0.20240226093414-135d7a718f4d
 	github.com/zitadel/schema v1.3.0
 	go.opentelemetry.io/otel v1.24.0
 	go.opentelemetry.io/otel/trace v1.24.0
-	golang.org/x/exp v0.0.0-20230817173708-d852ddb80c63
 	golang.org/x/oauth2 v0.17.0
 	golang.org/x/text v0.14.0
 )
